refactor: Move path formatting logic to a shared utility function
