+++ crates/forge_services/src/tools/fs/fs_read.rs
@@ -9,7 +9,7 @@ use schemars::JsonSchema;
 use serde::Deserialize;
 
 use crate::tools::utils::assert_absolute_path;
-use crate::{EnvironmentService, FsReadService, Infrastructure};
+use crate::{FsReadService, Infrastructure};
 
 #[derive(Deserialize, JsonSchema)]
 pub struct FSReadInput {
@@ -27,29 +27,6 @@ pub struct FSRead<F>(Arc<F>);
 impl<F: Infrastructure> FSRead<F> {
     pub fn new(f: Arc<F>) -> Self {
         Self(f)
-    }
-
-    /// Formats a path for display, converting absolute paths to relative when
-    /// possible
-    ///
-    /// If the path starts with the current working directory, returns a
-    /// relative path. Otherwise, returns the original absolute path.
-    fn format_display_path(&self, path: &Path) -> anyhow::Result<String> {
-        // Get the current working directory
-        let env = self.0.environment_service().get_environment();
-        let cwd = env.cwd.as_path();
-
-        // Try to create a relative path for display if possible
-        let display_path = if path.starts_with(cwd) {
-            match path.strip_prefix(cwd) {
-                Ok(rel_path) => rel_path.display().to_string(),
-                Err(_) => path.display().to_string(),
-            }
-        } else {
-            path.display().to_string()
-        };
-
-        Ok(display_path)
     }
 }
 
@@ -85,7 +62,8 @@ impl<F: Infrastructure> ExecutableTool for FSRead<F> {
 
         // Display a message about the file being read
         let title = "read";
-        let display_path = self.format_display_path(path)?;
+        let display_path =
+            crate::tools::fs::format_display_path(self.0.environment_service(), path)?;
         let message = TitleFormat::success(title).sub_title(display_path);
         println!("{}", message);
 
@@ -171,6 +149,7 @@ mod test {
             .to_string()
             .contains("Path must be absolute"));
     }
+
     #[tokio::test]
     async fn test_format_display_path() {
         let temp_dir = TempDir::new().unwrap();
@@ -178,10 +157,12 @@ mod test {
 
         // Create a mock infrastructure with controlled cwd
         let infra = Arc::new(MockInfrastructure::new());
-        let fs_read = FSRead::new(infra);
 
-        // Test with a mock path
-        let display_path = fs_read.format_display_path(Path::new(&file_path));
+        // Test with the common implementation
+        let display_path = crate::tools::fs::format_display_path(
+            infra.environment_service(),
+            Path::new(&file_path),
+        );
 
         // Since MockInfrastructure has a fixed cwd of "/test",
         // and our temp path won't start with that, we expect the full path
